@@ -1,5 +1,4 @@
 from pathlib import Path
-from typing import Optional
 
 import pandas as pd
 from pyiceberg.catalog import load_catalog
@@ -21,7 +20,7 @@ def create_sql_catalog(tmp_path: Path):
     return catalog
 
 
-def pyiceberg_to_pandas(table, sort_by=None, dtypes_like: Optional[pd.Series] = None):  # noqa: FA100
+def pyiceberg_to_pandas(table, sort_by=None, dtypes_like: pd.Series | None = None):
     df = table.scan().to_arrow().to_pandas()
     if sort_by is not None:
         if isinstance(sort_by, (list, tuple)):
