@@ -6,6 +6,8 @@ __all__ = [
     "SparkConnectServer",
 ]
 
+_native.initialize()
+
 
 class SparkConnectServer:
     """The Spark Connect server that uses Sail as the computation engine."""
@@ -18,10 +20,6 @@ class SparkConnectServer:
         :param port: The port to bind the server to.
         """
         self._inner = _native.spark.SparkConnectServer(ip, port)
-
-    def init_telemetry(self) -> None:
-        """Initialize OpenTelemetry for the server."""
-        self._inner.init_telemetry()
 
     def start(self, *, background=True) -> None:
         """Start the server.
